Copy files if they do not exist or if they were modified.

+++ uttl/buildout/copyfile/copyfile_recipe.py
@@ -25,45 +25,26 @@ class CopyFileRecipe(BaseRecipe):
 		if not 'files' in self.options:
 			raise UserError('Missing mandatory "files" option.')
 
-		self.files = self.options['files'].splitlines()
-		self.src_files = [os.path.join(self.src_path, file) for file in self.files]
-		self.dst_files = [os.path.join(self.dst_path, file) for file in self.files]
-
-		self.log.debug(str(self.files))
+		self.files = [os.path.join(self.dst_path, file) for file in self.options['files'].splitlines()]
 
 	def install(self):
-		installed = self.dst_files
-		invalid = []
-
-		# use private api to check for files that need to be installed
-
-		(installed_part_options, installed_exists) = self.buildout._read_installed_part_options()
-
-		if self.name in installed_part_options:
-			part_options = installed_part_options[self.name]
-			if part_options and '__buildout_installed__' in part_options:
-				installed = list(path.rstrip() for path in part_options['__buildout_installed__'].split())
-
-		self.log.debug('installed ' + str(installed))
-
-		invalid = [file for file in installed if not os.path.exists(file)]
-		self.log.debug('invalid ' + str(invalid))
-
-		for dst_path in invalid:
-			filename = os.path.basename(dst_path)
-			src_path = os.path.join(self.src_path, filename)
-			self.copyFile(src_path, dst_path, filename)
-
-		# check if files were modified
-
-		check_modified = list(set(invalid) - set(installed))
-		self.log.debug('check_modified ' + str(check_modified))
-		for dst_path in check_modified:
+		for dst_path in self.files:
 			filename = os.path.basename(dst_path)
 
 			src_path = os.path.join(self.src_path, filename)
+
+			# check if file is missing
+
+			if not os.path.exists(dst_path):
+				self.log.debug('%s does not exist at destination' % (filename))
+
+				self.copyFile(src_path, dst_path, filename)
+
+				continue
+
+			# check if source was modified
+
 			src_modified = os.path.getmtime(src_path)
-
 			dst_modified = os.path.getmtime(dst_path)
 
 			if src_modified > dst_modified:
@@ -71,31 +52,6 @@ class CopyFileRecipe(BaseRecipe):
 
 				self.copyFile(src_path, dst_path, filename)
 
-		"""
-		for file in invalid:
-			filename = os.path.basename(file)
-
-			src = os.path.abspath(file)
-			src_exists = os.path.exists(src)
-			if not src_exists:
-				src = os.path.abspath(os.path.join(self.src_path, filename))
-				src_exists = os.path.exists(src)
-
-			self.log.debug('src: %s (exists: %r)' % (src, src_exists))
-
-			if not src_exists:
-				raise FileNotFoundError(src)
-
-			dst = os.path.abspath(os.path.join(self.dst_path, filename))
-			dst_exists = os.path.exists(dst)
-
-			self.log.debug('dst: %s (exists: %r)' % (dst, dst_exists))
-
-			self.options.created(dst)
-
-			self.copyFile(src, dst)
-
-		"""
 		return self.options.created()
 
 	update = install
